+++ R/resamp_print.R
@@ -165,3 +165,7 @@ print.psa <- function(x, ...) {
   
   print(summary(x), ...)
 }
+
+get_frontier.psa <- function(x) {
+  get_frontier(get_model_results(x))
+}