+++ R/tabular_input.R
@@ -1373,7 +1373,7 @@ join_fits_across_time <- function(this_part) {
   if ("until" %in% names(this_part)) {
     this_part <- dplyr::arrange_(this_part, ~ until)
     
-    project_(dots = this_part$fit, 
+    join_(dots = this_part$fit, 
              at= this_part$until[!is.na(this_part$until)])
     
   } else {