+++ NEWS.md
@@ -1,5 +1,9 @@
 ## heemod 0.9.0.9001
 
+### Breaking changes
+
+  * Survival function renamed: `project()`=>`join()`.
+
 ### New features
 
   * NMB is now reportd as relative values.
@@ -7,6 +11,10 @@
 ### Bugfixes
 
   * Fixed error with `run_dsa()` in state expansion situations.
+
+### Backend changes
+
+  * Multiple changes to prepare for the `dplyr` update 0.6.0.
 
 ## heemod 0.9.0